feat: sort query

+++ src/app/services/employee.ts
@@ -1,7 +1,7 @@
 import { IEmployee } from "../interfaces/employees";
 import EmployeesRepository from "../repositories/employees";
-import Employee from "../models/employee";
 import AppError from "../errors/AppError";
+import { isValidObjectId } from "mongoose";
 
 class EmployeeService {
     newEmployeeRepository = new EmployeesRepository;
@@ -20,12 +20,17 @@ class EmployeeService {
         return await this.newEmployeeRepository.save(data);
     }
 
-    async findMany(): Promise<IEmployee[]> {
-        const employees =  await this.newEmployeeRepository.read()
-        return employees
+    async findMany(sorted: boolean | undefined): Promise<IEmployee[]> {
+        return await this.newEmployeeRepository.read(sorted)
     }
 
     async findOne(id: string): Promise<IEmployee | null> {
+        const objectId = isValidObjectId(id)
+
+        if(!objectId) {
+            throw new AppError("Invalid id", 400)
+        }
+
         const employee = await this.newEmployeeRepository.readOne(id)
 
         if(!employee) {
@@ -36,13 +41,38 @@ class EmployeeService {
     }
 
     async deleteOne(id: string): Promise<IEmployee | null> {
+        const objectId = isValidObjectId(id)
+
+        if(!objectId) {
+            throw new AppError("Invalid id", 400)
+        }
+
         const employee = await this.newEmployeeRepository.delete(id)
+
+        if(!employee) {
+            throw new AppError("Employee not found", 404)
+        }
 
         return employee
     }
 
     async update(id: string, data: IEmployee): Promise<IEmployee | null> {
+        const objectId = isValidObjectId(id)
+
+        if(!objectId) {
+            throw new AppError("Invalid id", 400)
+        }
+        
+        if (!data.actions || !data.charge || !data.department || !data.name) {
+            throw new Error("Todos os campos (actions, charge, department, name) são obrigatórios");
+        }
+
         const employee = await this.newEmployeeRepository.updateOne(id, data)
+        
+        if(!employee) {
+            throw new AppError("Employee not found", 404)
+        }
+        
         return employee
     }
 }
